Fixes Auto Scaling bug.
Setting the command.adjustentType when the viewState.operator changees.

When the Action operator changed it did not set it's value on the command.adjustmentType; which is what is sent in the POST to /task

This resulted in the task to always post the default "ChangeInCapacity" and now allow "ExactCapacity" to be set on the command object when the "Set to" option was chosen.