(core): upgrade to webpack 2
* Pulled `later` library in-house as it's abandoned and there's a bug in
the NPM's `index.js` file due to how the author was doing coverage
builds.  pulling in-house obviates the need for anything other than the
core lib.
* Happypack is not yet compatible with webpack 2 but the author is
looking into it.  See [this issue](https://github.com/amireh/happypack/issues/122)
for updates.  instead of removing all the happypack config for now, I've
left it in in the hopes it can one day be used again.
* The compile order of webpack config entries has changed so I had to
add a sort algorithm for the chunks for the HTML webpack plugin to
ensure the files were inserted into the generated `index.html` file in
the right order.
* Webpack 2 now requires the `-loader` suffix for all specified loaders,
so all references were updated to reflect this, however some loaders are
referenced by code outside the control of our config so I had to add the
`moduleExtension` option to fix this.  In theory, once those others are
updated to support Webpack 2, this can be removed.