Fast Property performance improvements
Reduces the number of watches on a the Propertie tab by using our good friend Mr. innerHtml to
render the list of properties.  This is necessary for applications that have a lot of properties. (I'm looking
at you API)

Also made some improvements around the filtering, and the filter tag generation.

Other small cosmetic tweeks:
- increased the max-size on the filter dropdown list
- Added spinners to indicate something was happening.