refactor(CI integration): Rename Jenkins to CI
CI integration is confusingly named Jenkins throughout Deck, even though
both Jenkins and Travis are supported, and more CI’s will probably be
supported in the future. This commit rename Jenkins to CI throughout the
UI. There are still some references to Jenkins left in some domain
objects (serverGroup.buildInfo.jenkins), but these will be addressed in
a separate PR.

BREAKING CHANGE: `jenkins` must be renamed to `ci` in the
`triggerTypes` section of `settings.js`. There’s also a new
property `feature.convertCiToJenkinsOnPipelineSave` that should
be set, but this will default to true to not break things.