feat(kubernetes): Allow Run Job trigger tag
Support resolving Run Job container tags from triggers. There are a
couple of things that can be improved here but this is basic
implementation. One thing to think about is what happens if there are 2
docker triggers and the pipeline is triggered by one that isn't used for
this job. What happens?

Changes include:

* Checkbox for selecting if you want to resolve from trigger
* If the box is checked, a list of Docker triggers is presented
* The Run Job Execution Details will not print the tag is it's from a
  trigger