fix(search): add default method value
* add default method value for post search result searchers so that it
can intelligently process search result map lookups that don't exist.
this is currently an artifact of region searching such that too many
search results are found and they're culled in clouddriver so that only
instance results come back.  we need to change the way the results are
returned for the new search to perhaps limit the search results to, say,
_n per category_ rather than _n total_.  need to have a think about
this.