fix(core): Link to child executions when no error messages available
No longer display the default `No reason provided.` error block.

spinnker/deck#4381 makes an attempt to link to other synthetic stages in
search of a suitable error message.

This PR goes a step further and attempts to link into a child execution
in the event that it was a pipeline or strategy that failed.