fix(provider/kubernetes): Disable editing of managed ReplicaSets
ReplicaSets owned by a Kubernetes controller should be scaled and
edited by using the server group manager to change the controller.

Otherwise the controller that owns the ReplicaSet will immediately
undo the scaling/editing.

Disable scaling and editing if the manifest is owned by a controller,
and include links to the server group manager if there is one.

Allow ReplicaSets to be deleted, but show a warning explaining that
the controller that owns the ReplicaSet may recreate it.

Fixes https://github.com/spinnaker/spinnaker/issues/2211