fix(provider/kubernetes): Show warnings for dashes in app names
Dashes in application names are allowed for applications created by
the Kubernetes v2 provider using moniker annotations.

Users should be allowed to create applications containing dashes when
using the Kubernetes provider, but be warned that such names should
only be used for v2.

In particular, this is important in order to be able to create the
application in deck and then provision templated pipelines using roer.