feat(provider/kubernetes): trim runJob logs from executions until explicitly requested
A Run Job's log output is no longer included as part of the pipeline
executions list response by default. The query param expand=true is
required to fetch the response including log output.

This PR adds that expand param to the executions service and updates
the k8s v1 runJob to only fetch the full log when a user clicks the
"Console Output" link.  The expand param currently goes unused.