feat(provider/kubernetes): add UID to instances
When instances are returned from Clouddriver via the server groups
or loadbalancer endpoints the serialized InstanceViewModel or
LoadBalancerInstance objects use instance's `name` as their `id`
in the response.

For Kubernetes instances the `name` is not globally unique.  This
resulted in UI problems such as clicking on one instance in the
infrastructure view leading to two instances being selected.

Clouddriver has been updated to return a `uid` field that is the
same as `id` for all providers except Kubernetes, which sets it
to the resource's UUID.  This new `uid` field needs to be used
in Deck wherever instance uniqueness is communicated or tested,
such as in URLs or during cluster filtering.