feat(authz/config): Add a pipeline roles section to pipeline config.
This feature allows users to select roles that are used for authorization
decisions during pipeline execution and removes the need for users to specify
a `runAsUser` field for each trigger. The feature is disabled by default and can
be enabled by setting `managedServiceAccountsEnabled` to true in `settings.js`.

Part of spinnaker/spinnaker#3037