fix(core/bootstrap): Remove vis=true/vis=false from URL after toggling visualizer
Adding ?vis=true to the end of the URL shows the visualizer.
However, many links in spinnaker will strip the vis=true from the URL, so the visualizer tends to disappear.
This change will enable/disable the visualizer when changing between `vis=true`/`vis=false`, but will then strip the parameter from the URL.
The visualizer will remain in its current state (enabled/disabled) until the parameter changes again.

This change exposed a bug in `@uirouter/core` which has also been fixed, but until the new version of ui-router is included in deck, the "remove parameter from the URL" behavior won't work reliably.