feat(pipeline_templates): Display "Force Rebake" checkbox for MPT
When manually executing templated pipelines, the "Force Rebake" checkbox is missing, even though the pipeline contains a bake stage. Templated pipelines do not contain any stages before runtime, so the bake stage is never asked to inject its custom logic into the manual execution modal.
This is a simple 90/10-solution that just iterates through previous executions, finds the previous execution that actually ran, and naïvely copies the stages to the pipeline, so that the bake stage magic is executed. It might result in false positives and negatives if you are changing the pipeline template, but as the main structure of a pipeline rarely changes and the fact that there are no bad consequences of displaying a force rebake checkbox when there is no bake stage, I think this is good enough.