refactor(artifacts): Clean up code to get kinds
All places where we call getKinds immediately filter on either
isMatch or isDefault; clean up the code an make the intention
more clear by having the API expose getMatchArtifactKinds and
getDefaultArtifactKinds.

Also, rename the defaultArtifactKind to customArtifactKind as
that fits more with how its usage is named; the fact that it is
the default in some cases is less fundamental than the fact that
it is custom.