refactor(artifacts): Generalize artifact delegate for reuse
The ongoing cloudformation stage effort will need to select an artifact
from the expected artifacts lists so it can be fetched during deploy,
quite similar to what the deploy manifest in kubernetes does.

This patch leverages the work being done for k8s, and generalizes the
manifest artifact delegate to be reused by the upcomming cloud formation.
It adds a tag attribute that will be prepended to ArtifactId and ArtifactAccount.
This way, different delegates can have its own attributes in the context without
conflicts.