feat(artifacts): Re-use artifacts when re-running a pipeline
The manual execution widget does not currently support injecting
artifacts into a pipeline. This means that pipelines that depend
on artifacts cannot be manually started; support this workflow
for the re-run pipeline case by re-sending the same artifacts as
the initial run. These will show up as a read-only list of artifacts
in the Confirm Execution dialog.