fix(artifacts): Simplify display name generation
The current logic to auto-generate a display name for artifacts
without one assumes that the user will enter a fully-qualified
path for each artifact type. If this is not true, an unhelpful
(and often duplicate) name such as GCS-no-bucket/no-file is
generated.

While we could make the logic more complex, it seems easier to
maintain and just as (if not more) useful to just default the
display name to the name.