fix(google): revert "select all zones by default when deploying a regional gce server group (#6751) (#6755)"
This reverts commit 63248626d8e416830b043fdc8db08d93e74e4bac.

In hindsight, Deck was already doing the right thing here. The platform, GCP,
chooses the zones if the user doesn't specify them. In some quick testing
through the GCP UI, MIGs in us-central1 always seem to get spread across
zones -b, -c, and -f and never end up on -a. So it was incorrect to show in
Deck that they might be placed on -a.