fix(artifacts): Allow defaulting in execution artifact
Find artifact from execution doesn't use the full expected artifact
component, which means users can't assign a name or use a default
artifact. Upgrade it to the full expected artifact component so
it gets these features.

Also fix the field defaulting; currently the fields will reset if
you navigate away and back to the stage before saving.

Also add the "Artifact Status" tab to this stage so it's easier
to see the artifact that was bound.