fix(runJob/kubernetes): use explicit pod name
the previous implementation of run job used manifest events to get the
name of the pod created by the job. this proved unreliable as events
roll off quickly in large environments. this commit implements a
`PodNameProvider` interface which supplies `JobManifestPodLogs` with the
name of the Pod to fetch logs from. Run Job supplies these Pod names by
collecting all Pods owned by the Job. For components which are still
event based, the `JobEventBasedPodNameProvider` implements the
previously used logic for getting the Pod name from the event.