fix(runJob/kubernetes): reliably display logs (#7060)
* fix(runJob/kubernetes): fix log display

fix a case where the log link was being displayed before all the data
needed to render it was available. this caused some confusing cases
where the link would be rendered but users would be met with a Not Found
error in the popup. ultimately this was due to a missing namespace.
resolved by pulling the namespace off of the job status instead of
depending on the manifest subscription.

* fix(runJob/kubernetes): fetch manifest once

jobs collect owned pods when collecting job status so we are no longer
dependent on subscribing to manifest events to get the pod names created
by the job. instead, simply fetch the manifest once. this is more
reliable than subscribing to manifest updates.