fix(k8s/runJob): null property file if value none
if the logs capture selector changes to `none` we should null out the
`propertyFile` field. the `consumeArtifactSource` field is used as a
flag for deck to determine which option is selected but it isn't taken
into account in orca. so, if `propertyFile` is non-null, orca will still
try to find properties even if we have `consumeArtifactSource = none`.

Fixes spinnaker/spinnaker#4642