feat(plugins): able to load plugin stages into UI
This adds the ability to add stages as plugins in the UI. There is an interface for plugin developers to use to create a UI for their stages. The stage UI can be built using ReactJS.

When Deck is starting up, it will look to see if any plugins are defined in `spinnakerSettings`. If there are plugins defined, a script tag will be appended to the bottom of the page and the plugin will have its initialized method called. The plugin initialize method takes in the `Registry` so the stage can register itself as a stage. Then the plugin developer has to call `window.spinnakerSettings.onPluginLoaded` with their plugin object.

We removed the `ng-app` from `index.deck` because of using `modules` in a promise. The bootstrapping of deck now happens in a separate Javascript file `app/scripts/bootstrap.js`. This script is added to the page via webpack. The reason for this change was tests were failing when calling `bootstrap`. `bootstrap` only happened before when loading the page by hitting `index.deck`. The tests were expecting things to not be loaded, but with `bootstrap` the entire application was loaded.

When loading plugins we didn't use CommonJS or ES modules due to keeping the list of supported browsers larger.