fix(core/pipeline): Don't break templated pipelines when updating config
When updatePipelineConfig is called on a templated pipeline, it breaks the UI rendering of the pipeline (since it is based on `$scope.plan`, not `$scope.pipeline`). This happens when you open the pipeline config screen for some pipelines (for instance if you have expected artifacts defined in the old format without artifact account).
Removing the line in this PR solves the problem, while it seems like the functionality still works (because angular watches `$scope.pipeline` and calls `updatePipeline()` when it changes).