fix(core/pipeline): Revision history is vertically challenged in Safari
Safari renders the viewport of the revision history with a height of 34px, which leaves a lot of whitespace unused, and also makes it hard to review the revision history as only two lines can be displayed at once.

The `padding-left` override also makes the revision history stop cropping the pluses and minuses (in all browsers).