feat(core/presentation): Begin adding support for error categories in validation API
This PR introduces initial support for two features:

- The ability to embed validation categories _inside_ a validation message and extract them later.
Validation categories are pre-defined by the validation API.
A validator can return a validation message prefixed with the category.
The `categorizeErrorMessage` function categorizes an error message into a pre-defined category.

Example:
the message 'Error: This is an error' is processed into:
category: 'error'
message: 'This is an error'.

- The ability to categorize validation messages from a raw Errors object (Formik style) into category buckets.
Given an errors object, the `categorizeErrors` function returns an object with category keys.
Each error message from the raw errors object is first categorized and then stored inside the categorized errors object.

Example:
This raw errors object: { foo: 'Warning: a warning', bar: 'Error: an error' }
is transformed into a categorized errors object:
{
  warning: { foo: 'a warning' },
  error: { bar: 'an error' } },
}

The plan is to then use the categorized error object within the SpinFormik component.

This approach (to embed the validation message category directly into the validation message) was chosen because it doesn't violate the contract of formik validation  That is, we still return strings from validation functions.  This will allow this feature to degrade gracefully, in a scenario where the outer code doesn't understand the categorized validation paradigm.