fix(core/instance): use fragment on non-UP health code path
Turns out that the `<OverlayTrigger>` component needs whatever you give it as children to be a single node. With the change introduced in #7520, the code path that runs when the health status is `UP` works fine, but any non-UP status hits the `<OverlayTrigger>` path and blows up/fails to render because there's two root nodes. This just wraps those two nodes in a fragment.