feat(core/presentation): Mark all SpinFormik initialValues fields as 'touched'
Normally, Formik fields do not show validation errors until the user has tabbed away (blur event).
This is to improve UX when a user first loads a blank form.
It would be horrible to see the entire form rendered in red because all the fields are required.

However, when a form is reloaded *with existing data*, we usually want to show validation errors immediately.

When the form is first rendered, mark all fields in initialValues as "touched".