feat(manualjudgment): Select roles that can execute manual judgement #4792
Added ability to add roles to manual judgment stage.
This is part of: spinnaker/spinnaker#4792.

Enhanced stage.html to

Get the roles of the application.
Display the roles of the application as a list only if the stage is Manual Judgment.
Enhanced stage.module.js to

Fetch the permissions of the application from the gate application url.
populate the list with only the roles of the application with no duplicates.
Enhanced ApplicationReader.ts to

Fetch the permissions of the application from the gate application url.
Enhanced ManualJudgmentApproval.tsx to

Fetch the roles of the application, stage and the user during the execution.
Iterate through each of the user role to check if the role exists in the stage and application.
If yes/no, enable/disable the continue button.
Display the instruction that User does not have permissions to continue