fix(angular-ui-bootstrap): Kludge: cast angular-ui-bootstrap default import as 'any'
This is due to the way the @types/angular-ui-bootstrap is structured.
There is no default import specified.
In a later version of the typings, there _is_ a default import declared, but the Interfaces are no longer exported (which we use).
Since we're not writing angularjs code anymore, I think this kludge is acceptable.

https://github.com/DefinitelyTyped/DefinitelyTyped/pull/37813#issuecomment-524466508