fix(artifacts): Fix find artifacts from execution stage
The find artifacts from execution stage is not able to load
the widget for each artifact type; creating the controller
is broken because the constructors are not annotated with
dependency injection information and we are running in strict-di
mode.

Fix this by extracting the anonmyous controller for each artifact
type into a function and annotating it with injection information.
(I'm not aware of a way to annotate an anonymous function, except
by doing Object.assign(function() {}, { $inject: 'abc' }) which
seemed less readable.)