refactor(core): clean up StageArtifactSelectorDelegate
Removes PreRewriteStageArtifactSelector and its corresponding props from StageArtifactSelectorDelegate. Adds a comment marking the delegate as deprecated, since now that there is only one UI for stage artifact selection, it is probably cleaner to just include that component inline in stage configs rather than adding an extra level of indirection that just wraps the component in a StageConfigField, which we're trying to migrate away from anyway.