refactor(halyard): remove defaultProviders from halconfig/settings
Halyard users are currently supplied with a list of defaultProviders that include every possible cloud provider. The AccountService reads from defaultProviders when trying to list all providers for a given application only if there the application does not have an explicit list of configured cloud providers, and then will take the intersection of available providers with these defaultProviders, but since we're supplying every possible provider in the list of defaultProviders, this is essentially the same as falling back to available providers, which is what we do when SETTINGS.defaultProviders is undefined.