fix(kubernetes): include manifest as $interpolation context for annotation-driven UI
The Kubernetes [annotation-driven UI](https://spinnaker.io/guides/user/kubernetes-v2/annotations-ui/) expects the full manifest to be available to hydrate custom UI sections specfied as annotations. Since we already have access to the full manifest in this component off of which we read each annotation, simply include it as part of the context we pass to $interpolate.