fix(amazon/rollback): Wait 500 ms before opening a second modal
When the react-bootstrap based confirmDialog closes, it removes modal-open class from the body tag after a 300ms fade out animation.  this causes the ui-bootstrap modal to not be styled properly.  Adding this delay allows the animation to complete and the modal-open class removed from the body tag.  Then, ui-bootstrap re-adds modal-open when it shows the rollback dialog.