refactor(kubernetes/pipelines): make ManifestStatus functional component
This component does not rely on any React lifecycle hooks, and we are encouraging all new contributions to prefer functional components. The @Overridable annotation only works on class components, and we are not sure if their were ever any custom extensions relying on it, so remove this for now. (Any custom extensions are unlikely to still work as expected anyway after the upcoming refactor in which data fetching is pushed down into each per-Manifest ManifestStatus component, so we can revisit appropriate extension points if anyone lets us know that this broke their custom overrides.)