refactor(kubernetes/pipelines): move translation from stage manifest to manifest params out of DeployStatus
Deck currently gets the full manifests it receives from Orca in the execution context, parses some details from those full manifests to create a new request to Clouddriver's /manifests endpoint. Much of this logic was added to disambiguate CRDs of the same kind but different API group. Ideally, Deck would have no knowledge of build-in Kubernetes groups and simply receive all the information it needs to construct a valid request. We'll revisit this later, but for now, let's at least move all the logic into private methods on the KubernetesManifestService rather than having them live on the DeployStatus component. This is also important because in a future commit, we will push the logic that sets up subscriptions against the /manifest endpoint further down the tree so that each ManifestStatus component is responsible for fetching its own data.