perf(kubernetes): Send server group and firewall requests in parallel
Similar to the change I just made for instances, let's request the
details and the manifest at the same time so that we reduce the amount
of time the user has to wait.

For server group managers and load balancers, we are not making a
request for the details at all, so we don't need to make the same
change to those controllers.