Added ability to add roles to manual judgment stage.
This is part of: spinnaker/spinnaker#4792.

Enhanced stage.html to

Get the roles of the application.
Display the roles of the application as a list, only if the stage is Manual Judgment.

Enhanced stage.module.js to

Fetch the permissions of the application from the gate application url.
populate the list with only the roles of the application with no duplicates.

Enhanced ApplicationReader.ts to

Fetch the permissions of the application from the gate application url.

Enhanced ManualJudgmentApproval.tsx to

Fetch the roles of the application, stage and the user during the execution.
Iterate through each of the user role to check if the role exists in the stage and application.
If the role exists in the stage roles added in the manual judgment stage,
then we check for role permissions added in the application.
If the stage role has 'READ', then we disable the continue button
If the stage role has 'WRITE,EXECUTE,CREATE', then we enable the continue button.

Enhanced help.contents.ts file

"The current user must have the permission to approve the manual judgment stage. Otherwise, you'll not be able continue to the next pipeline stage."