fix(kubernetes): Hide load balancers stage from Kubernetes provider
This stage is not supported for the Kubernetes provider; load balancers
should be created by using the 'Deploy (Manifest)' stage to deploy the
desired Kubernetes object (likely a service).

This PR just removes the 'Create Load Balancers' from the stage dropdown
for the Kubernetes provider to reduce confusion, as the stage doesn't
work for that provider.